feat: Task 모달에서 step분리
