+++ src/components/home/TaskModal/SelectRepeatModal.styled.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const SelectRepeatModalContent = styled.div`
+  grid-area: 2/2/3/3;
+  display: flex;
+`;
